common/bitutil: move fuzzer to other test packages
